+++ backend/hankkimoa/src/main/java/com/jjangiji/hankkimoa/auth/service/oauth/OauthRequestProperties.java
@@ -1,4 +1,4 @@
-package com.jjangiji.hankkimoa.auth.service;
+package com.jjangiji.hankkimoa.auth.service.oauth;
 
 import com.jjangiji.hankkimoa.auth.service.dto.request.OauthLoginRequest;
 import com.jjangiji.hankkimoa.common.exception.ExceptionCode;
